+++ crates/wedi-widget/src/layout/mod.rs
@@ -1,5 +1,0 @@
-pub mod line_layout;
-pub mod screen;
-
-pub use line_layout::LineLayout;
-pub use screen::ScreenLayout;